feat(admin): move dashboard to main
